// types.ts
export interface Database {
  public: {
    Tables: {
// ...
      orders: {
        Row: {
          id: string;
          created_at: string;
          user_id: string;
          order_items: CartItem[];
          first_name: string;
          last_name: string;
          address: string;
          city: string;
          state: string;
          postal_code: string;
          phone: string;
          email: string;
          total: number;
          subtotal: number;
          shipping: number;
          paid: boolean;
          payment_id: string;
          paid_at: string;
        };
        Insert: {
          id?: string;
          created_at?: string;
          user_id?: string;
          order_items?: CartItem[];
          first_name?: string;
          last_name?: string;
          address?: string;
          city?: string;
          state?: string;
          postal_code?: string;
          phone?: string;
          email?: string;
          total?: number;
          subtotal?: number;
          shipping?: number;
          paid?: boolean;
          payment_id?: string;
          paid_at?: string;
        };
        Update: {
          id?: string;
          created_at?: string;
          user_id?: string;
          order_items?: CartItem[];
          first_name?: string;
          last_name?: string;
          address?: string;
          city?: string;
          state?: string;
          postal_code?: string;
          phone?: string;
          email?: string;
          total?: number;
          subtotal?: number;
          shipping?: number;
          paid?: boolean;
          payment_id?: string;
          paid_at?: string;
        };
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      get_total_paid_orders: {
        Args: Record<PropertyKey, never>;
        Returns: number;
      };
      is_admin: {
        Args: Record<PropertyKey, never>;
        Returns: boolean;
      };
    };
    Enums: {
      [_ in never]: never;
    };
  };
}